fix: resolved imports & fixed structure
