+++ apps/backoffice-v2/src/domains/checks/hooks/queries/useIdentityVerificationChecksQuery/useIdentityVerificationChecksQuery.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 
 import { useIsAuthenticated } from '@/domains/auth/context/AuthProvider/hooks/useIsAuthenticated/useIsAuthenticated';
 import { IIdentityVerificationChecksParams } from '@/domains/identity-verification/fetchers';
-import { identityVerificationChecksQueryKey } from '@/domains/identity-verification/query-keys';
+import { identityVerificationChecksQueryKey } from '@/domains/checks/query-keys';
 
 export const useIdentityVerificationChecksQuery = ({
   ...params
